Added clearTextAndTokens()
clearTextAndTokens() does what I wanted clear() to do, which is to remove all objects and stray text from a TokenCompleteTextView. The problem is that clear() removes spans instead of objects.
I was unable to safely remove stray text from the view after invoking removeObject() for all objects because it posts a Runnable instead of immediately removing the object.