Don't hook into invalidate() to trigger redrawing spans
invalidate() often triggers additional invalidate() calls and it's easy to get into an endless loop or do too much work. Doing additional work in invalidate() is also the cause of a crash when cutting all text.
Instead of using an implementation detail of TextView to trigger redrawing the spans we now add or remove an invisible dummy span to achieve the same result. This is very unlikely to break in the future.