Add work-around for API 26-29 hint failure on first completion
With an empty text field with a hint and prefix using the stock emulator
keyboard, the first completion will not complete and instead erase all user
input. This work-around postpones updating the hint until after the whole
completion process is finished to prevent the hint from overwriting this first
set of input. Without this, you would type "ma," in the sample project and
everything gets erased